+++ globalConfig.json
@@ -1,1 +1,1 @@
-{"mongoUri":"mongodb://127.0.0.1:54907/","mongoDBName":"jest"}
+{"mongoUri":"mongodb://127.0.0.1:64224/","mongoDBName":"jest"}
