+++ src/main/factories/controllers/login/login-controller-factory.ts
@@ -1,6 +1,6 @@
 import { makeLogControllerDecorator } from '@main/factories/decorators/log-controller-decorator-factory'
 import { makeDbAuthentication } from '@main/factories/usecases/authentication/db-authentication-factory'
-import { LoginController } from '@presentation/controller/login/login-controller'
+import { LoginController } from '@presentation/controller/auth/login/login-controller'
 import { Controller } from '@presentation/protocols'
 import { makeLoginValidation } from './login-validation-factory'
