+++ src/main/factories/controllers/signup/signup-controller-factory.ts
@@ -1,7 +1,7 @@
 import { makeLogControllerDecorator } from '@main/factories/decorators/log-controller-decorator-factory'
 import { makeDbAddAccount } from '@main/factories/usecases/add-account/db-add-account-factory'
 import { makeDbAuthentication } from '@main/factories/usecases/authentication/db-authentication-factory'
-import { SignUpController } from '@presentation/controller/signup/signup-controller'
+import { SignUpController } from '@presentation/controller/auth/signup/signup-controller'
 import { Controller } from '@presentation/protocols'
 import { makeSignUpValidation } from './signup-validation-factory'
