+++ tsconfig.json
@@ -23,5 +23,5 @@
     "rootDir": "src"
   },
   "include": ["src"],
-  "exclude": ["**/*.spec.ts", "**/*.test.ts"]
+  // "exclude": ["**/*.spec.ts", "**/*.test.ts"]
 }
